Extract e2e setup helpers
