+++ frontend/test/__support__/e2e/helpers/e2e-setup-helpers.js
@@ -1,0 +1,8 @@
+export function snapshot(name) {
+  cy.request("POST", `/api/testing/snapshot/${name}`);
+}
+
+export function restore(name = "default") {
+  cy.log("Restore Data Set");
+  cy.request("POST", `/api/testing/restore/${name}`);
+}
